pyproject.toml
[tool.poetry]
name = "openlift"
version = "0.1.0"
description = "Marketing incrementality measurement using geo-lift with Bayesian counterfactuals in PyMC"
authors = ["OpenLift Team <team@openlift.ai>"]
readme = "README.md"
packages = [{include = "openlift"}]

[tool.poetry.dependencies]
python = "~3.11"
pymc = "^5.10.0"
arviz = "^0.19.0"
pandas = "^2.1.0"
numpy = "<2.0.0" 
typer = {extras = ["all"], version = "^0.15.0"}
pydantic = "^2.5.0"
ruamel-yaml = "^0.18.0"
streamlit = "^1.31.0"
matplotlib = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.1.0"
black = "^24.0.0"
mypy = "^1.8.0"
mkdocs = "^1.5.0"

[tool.poetry.scripts]
openlift = "openlift.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']
